:art: fix: split models(event, slack), routers(event) & rename middleware, routers, models files #90
