+++ backend/app.ts
@@ -6,10 +6,10 @@ import swaggerUI from "swagger-ui-express";
 import cookieParser from "cookie-parser";
 import expressSession from "express-session";
 
-import { router } from "./controllers/routers/route";
-import { connectionForCabinet } from "./models/db";
-import passportConfig from "./controllers/middleware/passport";
-import slack from "./controllers/middleware/slack";
+import { router } from "./controllers/routers/authRouter";
+import { connectionForCabinet } from "./models/dbModel";
+import passportConfig from "./controllers/middleware/passportMiddleware";
+import slack from "./controllers/middleware/slackMiddleware";
 
 import dotenv from "dotenv";
