+++ backend/controllers/middleware/slackMiddleware.ts
@@ -1,8 +1,8 @@
 import schedule from "node-schedule";
 import { WebClient } from "@slack/web-api";
-import { slackReturnUser, slackOverdueUser } from "../../models/query";
+import { slackReturnUser, slackOverdueUser } from "../../models/slackModel";
 import dotenv from "dotenv";
-import { slackUserList, slackUser } from "../../models/user";
+import { slackUserList, slackUser } from "../../models/userModel";
 
 const env = process.env;
 if (env.USER === "ec2-user") {
