+++ backend/models/queryModel.ts
@@ -1,9 +1,8 @@
 import mariadb from "mariadb";
-import { lentInfo, lentCabinetInfo, userInfo } from "./user";
-import { sendLentMsg, sendReturnMsg } from "../controllers/middleware/slack";
-import { eventInfo } from "./user";
+import { lentInfo, lentCabinetInfo, userInfo } from "./userModel";
+import { sendLentMsg, sendReturnMsg } from "../controllers/middleware/slackMiddleware";
 
-const con = mariadb.createPool({
+export const con = mariadb.createPool({
   host: "localhost",
   user: "root",
   password: "",
@@ -214,136 +213,4 @@ export async function activateExtension(user: any) {
       throw err;
     });
   if (pool) pool.end();
-}
-
-//슬랙 연체 메세지 발송을 위한 연체 유저 정보 조회
-export async function slackOverdueUser(day: string) {
-  let pool: mariadb.PoolConnection;
-  let intraList: Array<string> = [];
-  const content: string = `SELECT intra_id FROM user INNER JOIN lent ON lent.lent_user_id = user.user_id WHERE expire_time='${day}'`;
-  pool = await con.getConnection();
-  await pool
-    .query(content)
-    .then((res: any) => {
-      if (res) {
-        for (let i = 0; i < res.length; i++) {
-          intraList.push(res[i].intra_id);
-        }
-      }
-    })
-    .catch((err: any) => {
-      console.log(err);
-      throw err;
-    });
-  if (pool) pool.end();
-  return intraList;
-}
-
-//슬랙 메세지 발송을 위한 반납 전날 유저 정보 조회
-export async function slackReturnUser(day: string) {
-  let pool: mariadb.PoolConnection;
-  let intraList: Array<string> = [];
-  const content: string = `SELECT intra_id FROM user INNER JOIN lent ON lent.lent_user_id = user.user_id WHERE expire_time='${day}'`;
-  pool = await con.getConnection();
-  await pool
-    .query(content)
-    .then((res: any) => {
-      if (res) {
-        for (let i = 0; i < res.length; i++) {
-          intraList.push(res[i].intra_id);
-        }
-      }
-    })
-    .catch((err: any) => {
-      console.log(err);
-      throw err;
-    });
-  if (pool) pool.end();
-  return intraList;
-}
-
-// event 정보 조회
-export async function getEventInfo(intra_id: string) {
-  let pool: mariadb.PoolConnection;
-  let eventInfo: Array<eventInfo> = [];
-  const selectContent: string = `select b.* from (select * from event where intra_id="${intra_id}") as a, event as b where b.event_id = a.event_id or b.event_id = a.event_id + if(a.event_id % 2 = 0, - 1, + 1)`;
-
-  pool = await con.getConnection();
-  await pool
-    .query(selectContent)
-    .then((res: any) => {
-      eventInfo = res;
-    })
-    .catch((err: any) => {
-      console.log(err);
-      throw err;
-    });
-  if (pool) pool.end();
-  return ({eventInfo: eventInfo});
-}
-
-// event 정보 추가
-export async function insertEventInfo(intra_id:string) {
-  let pool: mariadb.PoolConnection;
-  const checkContent: string = `select count(*) as count from event where intra_id="${intra_id}"`;
-  const insertContent: string = `update event as a, (select * from event where isEvent=false limit 1) as b set a.isEvent=true, a.intra_id="${intra_id}" where a.event_id = b.event_id`;
-
-  pool = await con.getConnection();
-  await pool
-  .query(checkContent)
-  .then(async(res:any) => {
-    if (res[0].count === 0)
-    {
-      await pool
-        .query(insertContent)
-        .then((result:any) => {})
-        .catch((err:any) => {
-          console.log(err);
-          throw err;
-        });
-    }
-    else {
-      console.log(`${intra_id} : 이미 이벤트 정보가 있습니다.`);
-    }
-  })
-  .catch((err: any) => {
-    console.log(err);
-    throw err;
-  })
-
-  if (pool) pool.end();
-  return ({ errno: 0 });
-}
-
-// event 정보 update
-export async function updateEventInfo(intra_id:string) {
-  let pool: mariadb.PoolConnection;
-  const checkContent: string = `select count(*) as count from event where intra_id="${intra_id}"`;
-  const updateContent: string = `update 42cabi_DB.event set intra_id = NULL, isEvent=false where intra_id="${intra_id}"`;
-
-  pool = await con.getConnection();
-  await pool
-  .query(checkContent)
-  .then(async(res:any) => {
-    if (res[0].count > 0)
-    {
-      await pool
-        .query(updateContent)
-        .then((result:any) => {})
-        .catch((err:any) => {
-          console.log(err);
-          throw err;
-        });
-    }
-    else {
-      console.log(`${intra_id} : 이벤트 대상자가 아닙니다.`);
-    }
-  })
-  .catch((err: any) => {
-    console.log(err);
-    throw err;
-  })
-
-  if (pool) pool.end();
-  return ({ errno: 0 });
 }
